+++ migrations/mediadb/sqlite/004-add-io_state-for-image_object.sql
@@ -1,0 +1,5 @@
+/* #############
+# add io_state to image_object
+############# */
+ALTER TABLE image_object ADD COLUMN io_state VARCHAR(50) DEFAULT 'UNKNOWN';
+